+++ MANIFEST.in
@@ -2,5 +2,5 @@ include LICENSE
 include MANIFEST.in
 include versioneer.py
 
-graft arsenic
+graft cinnabar
 global-exclude *.py[cod] __pycache__ *.so
