+++ cinnabar/plotting.py
