+++ cinnabar/stats.py
