+++ cinnabar/tests/conftest.py
@@ -1,14 +1,14 @@
 import pytest
 from importlib import resources
 
-from arsenic.wrangle import FEMap
+from cinnabar.wrangle import FEMap
 
 
 @pytest.fixture(scope="session")
 def fe_map():
     """FEMap using test csv data"""
 
-    with resources.path("arsenic.data", "example.csv") as fn:
+    with resources.path("cinnabar.data", "example.csv") as fn:
         femap = FEMap(fn)
 
     return femap
