+++ devtools/conda-envs/env.yaml
@@ -1,4 +1,4 @@
-name: arsenic
+name: cinnabar
 channels:
   - conda-forge
 dependencies:
