+++ docs/changelog.rst
@@ -6,11 +6,11 @@ Release history
 
 This section lists features and improvements of note in each release.
 
-The full release history can be viewed `at the GitHub arsenic releases page <https://github.com/OpenFreeEnergy/arsenic/releases>`_.
+The full release history can be viewed `at the GitHub cinnabar releases page <https://github.com/OpenFreeEnergy/cinnabar/releases>`_.
 
 0.2.1 - Release
 ---------------
 
 Bugfixes
 ^^^^^^^^
-Fix erroneous MLE estimate when self-edges are included (PR `#38 <https://github.com/openforcefield/openff-arsenic/pull/38>`_).
+Fix erroneous MLE estimate when self-edges are included (PR `#38 <https://github.com/OpenFreeEnergy/cinnabar/pull/38>`_).
