+++ docs/getting_started.rst
@@ -1,4 +1,4 @@
 Getting Started
 ===============
 
-`arsenic` is a package to aid with the consistent and reliable plotting of results from free energy calculations. Fair assessment and comparison of computational methods relies on robust and reproduceable metrics, which `freeenergy framework` aims to provide.
+`cinnabar` is a package to aid with the consistent and reliable plotting of results from free energy calculations. Fair assessment and comparison of computational methods relies on robust and reproduceable metrics, which `freeenergy framework` aims to provide.
