+++ docs/index.rst
@@ -1,9 +1,9 @@
-.. arsenic documentation master file, created by
+.. cinnabar documentation master file, created by
    sphinx-quickstart on Thu Mar 15 13:55:56 2018.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to arsenic's documentation!
+Welcome to cinnabar's documentation!
 =========================================================
 
 .. toctree::
