+++ setup.cfg
@@ -4,7 +4,7 @@
 # .coveragerc to control coverage.py and pytest-cov
 omit =
     # Omit generated versioneer
-    arsenic/_version.py
+    cinnabar/_version.py
 branch = True
 
 [yapf]
@@ -21,10 +21,10 @@ max-line-length = 119
 # Automatic version numbering scheme
 VCS = git
 style = pep440
-versionfile_source = arsenic/_version.py
-versionfile_build = arsenic/_version.py
+versionfile_source = cinnabar/_version.py
+versionfile_build = cinnabar/_version.py
 tag_prefix = ''
-parentdir_prefix = arsenic-
+parentdir_prefix = cinnabar-
 
 [aliases]
 test = pytest
@@ -32,4 +32,4 @@ test = pytest
 include_package_data = True
 
 [options.package_data]
-arsenic = "./arsenic/data/example.csv"
+cinnabar = "./cinnabar/data/example.csv"
